bug fixes with directory references
