+++ lib/engineer.js
@@ -20,7 +20,7 @@ class Engineer extends Employee {
         return 'Engineer';
     }
     generateCard = () => {
-        fs.appendFile('../dist/index.html',
+        fs.appendFile('./dist/index.html',
             `
                     <div class="engineer card bg-secondary" style="width: 18rem; margin: 20px;">
                         <div class="card-body">
@@ -37,7 +37,7 @@ class Engineer extends Employee {
             (err) => {
                 if (err)
                     throw err;
-                console.log(`Successfully added to index.html!\n`);
+                console.log(`\n----------------------------\nSuccessfully added to index.html!`);
             });
     }
 }
