+++ lib/intern.js
@@ -21,7 +21,7 @@ class Intern extends Employee {
     }
 
     generateCard = () => {
-        fs.appendFile('../dist/index.html',
+        fs.appendFile('./dist/index.html',
             `
                     <div class="intern card bg-secondary" style="width: 18rem; margin: 20px;">
                         <div class="card-body">
@@ -38,7 +38,7 @@ class Intern extends Employee {
             (err) => {
                 if (err)
                     throw err;
-                console.log(`Successfully added to index.html!\n`);
+                console.log(`\n----------------------------\nSuccessfully added to index.html!`);
             });
     }
 }
