+++ lib/manager.js
@@ -21,7 +21,7 @@ class Manager extends Employee {
     }
 
     generateCard = () => {
-        fs.appendFile('../dist/index.html',
+        fs.appendFile('./dist/index.html',
             `
                     <div class="manager card bg-secondary" style="width: 18rem; margin: 20px;">
                         <div class="card-body">
@@ -38,7 +38,7 @@ class Manager extends Employee {
             (err) => {
                 if (err)
                     throw err;
-                console.log(`Successfully added to index.html!\n`);
+                console.log(`\n----------------------------\nSuccessfully added to index.html!`);
             });
     }
 }
